+++ crates/primitives/src/integer/integer64/value.rs
@@ -5,8 +5,9 @@ pub(crate) fn get(value: &Value) -> Result<i64, CoreError> {
     value
         .downcast_ref::<i64>()
         .copied()
-        .ok_or_else(|| CoreError::InvalidValueRepresentation("int".into()))
+        .ok_or_else(|| CoreError::InvalidValueRepresentation("int64".into()))
 }
+
 #[cfg(test)]
 #[path = "value.tests.rs"]
 mod tests;
